refactor(all): package 구조 변경
