+++ src/main/java/lotto/Application.java
@@ -2,7 +2,7 @@ package lotto;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.view.InputView;
-import lotto.view.controller.LottoController;
+import lotto.controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
