+++ src/main/java/lotto/domain/Lotto.java
@@ -1,10 +1,10 @@
-package lotto.view.domain;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.view.global.exception.CustomException;
-import lotto.view.global.exception.ErrorMessage;
+import lotto.global.exception.CustomException;
+import lotto.global.exception.ErrorMessage;
 
 public class Lotto {
     private static final int MIN_NUMBER = 1;
