+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.view.domain;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
