+++ src/main/java/lotto/domain/WinningCondition.java
@@ -1,4 +1,4 @@
-package lotto.view.domain;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
