+++ src/main/java/lotto/global/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto.view.global.exception;
+package lotto.global.exception;
 
 public enum ErrorMessage {
     BLANK_INPUT("빈 문자열이 입력되었습니다."),
