+++ src/test/java/lotto/domain/WinningResultTest.java
@@ -4,11 +4,6 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueN
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.view.domain.Amount;
-import lotto.view.domain.Lotto;
-import lotto.view.domain.Lottos;
-import lotto.view.domain.WinningCondition;
-import lotto.view.domain.WinningResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
